test5_turborag_CAG/core/requirements.txt
docling >=0.1.3
faiss-cpu >=1.8.0
pdfminer.six
ofxparse
streamlit>=1.32.0
PyPDF2>=3.0.1
requests>=2.31.0
numpy>=1.26.4
sentence-transformers>=2.2.2
langchain>=0.1.16
langchain-docling>=0.1.3
pandas>=2.1.3
plotly>=5.17.0
schedule>=1.2.0